@@ -4,8 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-//go:build (solaris && !cgo) || (darwin && !cgo) || (android && amd64)
-// +build solaris,!cgo darwin,!cgo android,amd64
+//go:build (solaris && !cgo) || (darwin && !cgo) || (android && amd64) || ios || (darwin && kqueue)
+// +build solaris,!cgo darwin,!cgo android,amd64 ios darwin,kqueue
 
 package fs
 
