@@ -4,8 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//go:build (!windows && !linux) || android
-// +build !windows,!linux android
+//go:build (!windows && !linux && !ios) || android
+// +build !windows,!linux,!ios android
 
 package osutil
 
