@@ -4,8 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//go:build noupgrade
-// +build noupgrade
+//go:build noupgrade || ios
+// +build noupgrade ios
 
 package upgrade
 
