@@ -75,6 +75,7 @@ type FolderConfiguration struct {
 	RescanIntervalS         int                         `protobuf:"varint,7,opt,name=rescan_interval_s,json=rescanIntervalS,proto3,casttype=int" json:"rescanIntervalS" xml:"rescanIntervalS,attr" default:"3600"`
 	FSWatcherEnabled        bool                        `protobuf:"varint,8,opt,name=fs_watcher_enabled,json=fsWatcherEnabled,proto3" json:"fsWatcherEnabled" xml:"fsWatcherEnabled,attr" default:"true"`
 	FSWatcherDelayS         float64                     `protobuf:"fixed64,9,opt,name=fs_watcher_delay_s,json=fsWatcherDelayS,proto3" json:"fsWatcherDelayS" xml:"fsWatcherDelayS,attr" default:"10"`
+	FSWatcherTimeoutS       float64                     `protobuf:"fixed64,40,opt,name=fs_watcher_timeout_s,json=fsWatcherTimeoutS,proto3" json:"fsWatcherTimeoutS" xml:"fsWatcherTimeoutS,attr"`
 	IgnorePerms             bool                        `protobuf:"varint,10,opt,name=ignore_perms,json=ignorePerms,proto3" json:"ignorePerms" xml:"ignorePerms,attr"`
 	AutoNormalize           bool                        `protobuf:"varint,11,opt,name=auto_normalize,json=autoNormalize,proto3" json:"autoNormalize" xml:"autoNormalize,attr" default:"true"`
 	MinDiskFree             Size                        `protobuf:"bytes,12,opt,name=min_disk_free,json=minDiskFree,proto3" json:"minDiskFree" xml:"minDiskFree" default:"1 %"`
@@ -240,158 +241,161 @@ func init() {
 }
 
 var fileDescriptor_44a9785876ed3afa = []byte{
-	// 2413 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x58, 0xcf, 0x6f, 0x24, 0x47,
-	0xf5, 0x77, 0xdb, 0xbb, 0x6b, 0xbb, 0xfc, 0xbb, 0x6c, 0xef, 0x76, 0x9c, 0xc4, 0x35, 0xe9, 0xcc,
-	0x26, 0x4e, 0xbe, 0x89, 0x77, 0xe3, 0x44, 0x91, 0x12, 0x7d, 0x03, 0x64, 0xec, 0x8c, 0x58, 0x16,
-	0x67, 0xad, 0x1e, 0x43, 0x20, 0x41, 0x6a, 0xda, 0xdd, 0x35, 0x33, 0x1d, 0xf7, 0x8f, 0xa1, 0xab,
-	0xbd, 0xf6, 0xec, 0x21, 0x0a, 0x39, 0x20, 0x24, 0x72, 0x40, 0xe6, 0x80, 0x38, 0x20, 0x45, 0x02,
-	0x21, 0x08, 0x17, 0xce, 0xfc, 0x05, 0xb9, 0x20, 0xfb, 0x84, 0x10, 0x87, 0x92, 0xe2, 0xbd, 0xcd,
-	0x71, 0x8e, 0x7b, 0x42, 0xef, 0xf5, 0x8f, 0xa9, 0x9e, 0x99, 0x48, 0x48, 0xdc, 0xa6, 0x3e, 0x9f,
-	0x57, 0xef, 0x7d, 0xfa, 0x55, 0xd5, 0xab, 0x57, 0x43, 0xaa, 0xbe, 0x77, 0x74, 0xc7, 0x89, 0xc2,
-	0xa6, 0xd7, 0xba, 0xd3, 0x8c, 0x7c, 0x97, 0xc7, 0xe9, 0xe0, 0x24, 0xb6, 0x13, 0x2f, 0x0a, 0xb7,
-	0x3b, 0x71, 0x94, 0x44, 0xf4, 0x46, 0x0a, 0x6e, 0x3c, 0x3d, 0x62, 0x9d, 0x74, 0x3b, 0x3c, 0x35,
-	0xda, 0x58, 0x57, 0x48, 0xe1, 0x3d, 0xca, 0xe1, 0x0d, 0x05, 0xee, 0x9c, 0xf8, 0x7e, 0x14, 0xbb,
-	0x3c, 0xce, 0xb8, 0x2d, 0x85, 0x7b, 0xc8, 0x63, 0xe1, 0x45, 0xa1, 0x17, 0xb6, 0xc6, 0x28, 0xd8,
-	0x60, 0x8a, 0xe5, 0x91, 0x1f, 0x39, 0xc7, 0xc3, 0xae, 0x28, 0x18, 0x34, 0xc5, 0x1d, 0x10, 0x24,
-	0x32, 0xec, 0x99, 0x0c, 0x73, 0xa2, 0x4e, 0x37, 0xb6, 0xc3, 0x16, 0x0f, 0x78, 0xd2, 0x8e, 0xdc,
-	0x8c, 0x9d, 0xe5, 0x67, 0x49, 0xfa, 0xd3, 0xf8, 0xe7, 0x14, 0x79, 0xaa, 0x8e, 0xdf, 0xb3, 0xc7,
-	0x1f, 0x7a, 0x0e, 0xdf, 0x55, 0x15, 0xd0, 0x2f, 0x35, 0x32, 0xeb, 0x22, 0x6e, 0x79, 0xae, 0xae,
-	0x55, 0xb4, 0xad, 0xf9, 0xda, 0xe7, 0xda, 0x57, 0x92, 0x4d, 0xfc, 0x5b, 0xb2, 0x37, 0x5a, 0x5e,
-	0xd2, 0x3e, 0x39, 0xda, 0x76, 0xa2, 0xe0, 0x8e, 0xe8, 0x86, 0x4e, 0xd2, 0xf6, 0xc2, 0x96, 0xf2,
-	0x0b, 0x24, 0x60, 0x10, 0x27, 0xf2, 0xb7, 0x53, 0xef, 0xf7, 0xf6, 0xae, 0x24, 0x9b, 0xc9, 0x7f,
-	0xf7, 0x24, 0x9b, 0x71, 0xb3, 0xdf, 0x7d, 0xc9, 0x16, 0xce, 0x02, 0xff, 0x6d, 0xc3, 0x73, 0x5f,
-	0xb1, 0x93, 0x24, 0x36, 0x7a, 0x17, 0xd5, 0xe9, 0xec, 0x77, 0xff, 0xa2, 0x5a, 0xd8, 0xfd, 0xf2,
-	0xb2, 0xaa, 0x9d, 0x5f, 0x56, 0x0b, 0x1f, 0x66, 0xce, 0xb8, 0xf4, 0x4f, 0x1a, 0x59, 0xf0, 0xc2,
-	0x24, 0x8e, 0xdc, 0x13, 0x87, 0xbb, 0xd6, 0x51, 0x57, 0x9f, 0x44, 0xc1, 0x9f, 0xfe, 0x4f, 0x82,
-	0x7b, 0x92, 0xcd, 0x0f, 0xbc, 0xd6, 0xba, 0x7d, 0xc9, 0x6e, 0xa5, 0x42, 0x15, 0xb0, 0x90, 0xbc,
-	0x32, 0x82, 0x82, 0x60, 0xb3, 0xe4, 0x81, 0x3a, 0x64, 0x95, 0x87, 0x4e, 0xdc, 0xed, 0x40, 0x8e,
-	0xad, 0x8e, 0x2d, 0xc4, 0x69, 0x14, 0xbb, 0xfa, 0x54, 0x45, 0xdb, 0x9a, 0xad, 0xed, 0xf4, 0x24,
-	0xa3, 0x03, 0xfa, 0x20, 0x63, 0xfb, 0x92, 0xe9, 0x18, 0x76, 0x94, 0x32, 0xcc, 0x31, 0xf6, 0xc6,
-	0x79, 0x95, 0xac, 0xa6, 0x0b, 0x5b, 0x5e, 0xd2, 0x06, 0x99, 0xcc, 0x96, 0x72, 0xb6, 0xb6, 0x7b,
-	0x25, 0xd9, 0x24, 0x7e, 0xe2, 0xa4, 0x07, 0x11, 0x36, 0x4b, 0x2b, 0x50, 0x09, 0x23, 0x97, 0x37,
-	0xed, 0x13, 0x3f, 0x79, 0xdb, 0x48, 0xe2, 0x13, 0xae, 0x2e, 0xc9, 0xf9, 0x65, 0x75, 0xf2, 0xde,
-	0xde, 0x17, 0xf0, 0x6d, 0x93, 0x9e, 0x4b, 0x7f, 0x40, 0xae, 0xfb, 0xf6, 0x11, 0xf7, 0x31, 0xe3,
-	0xb3, 0xb5, 0x6f, 0xf7, 0x24, 0x4b, 0x81, 0xbe, 0x64, 0x15, 0x74, 0x8a, 0xa3, 0xcc, 0x6f, 0xcc,
-	0x45, 0x62, 0xc7, 0xc9, 0xdb, 0x46, 0xd3, 0xf6, 0x05, 0xba, 0x25, 0x03, 0xfa, 0xd3, 0xcb, 0xea,
-	0x84, 0x99, 0x4e, 0xa6, 0x2d, 0xb2, 0xd4, 0xf4, 0x7c, 0x2e, 0xba, 0x22, 0xe1, 0x81, 0x05, 0xfb,
-	0x1b, 0x93, 0xb4, 0xb8, 0x43, 0xb7, 0x9b, 0x62, 0xbb, 0x5e, 0x50, 0x87, 0xdd, 0x0e, 0xaf, 0xbd,
-	0xdc, 0x93, 0x6c, 0xb1, 0x59, 0xc2, 0xfa, 0x92, 0xad, 0x61, 0xf4, 0x32, 0x6c, 0x98, 0x43, 0x76,
-	0x74, 0x9f, 0x5c, 0xeb, 0xd8, 0x49, 0x5b, 0xbf, 0x86, 0xf2, 0xdf, 0xea, 0x49, 0x86, 0xe3, 0xbe,
-	0x64, 0x4f, 0xe3, 0x7c, 0x18, 0x64, 0xe2, 0x8b, 0x94, 0x7c, 0x02, 0xc2, 0x67, 0x0b, 0xe6, 0xc9,
-	0x45, 0x55, 0xfb, 0xc4, 0xc4, 0x69, 0xf4, 0x80, 0x5c, 0x43, 0xb1, 0xd7, 0x33, 0xb1, 0xe9, 0xe9,
-	0xdd, 0x4e, 0x97, 0x03, 0xc5, 0x6e, 0x41, 0x88, 0x24, 0x95, 0xb8, 0x84, 0x21, 0x60, 0x50, 0x6c,
-	0xa3, 0xd9, 0x62, 0x64, 0xa2, 0x15, 0xfd, 0x09, 0x99, 0x4e, 0xf7, 0xb9, 0xd0, 0x6f, 0x54, 0xa6,
-	0xb6, 0xe6, 0x76, 0x9e, 0x2b, 0x3b, 0x1d, 0x73, 0x78, 0x6b, 0x0c, 0xb6, 0x7d, 0x4f, 0xb2, 0x7c,
-	0x66, 0x5f, 0xb2, 0x79, 0x0c, 0x95, 0x8e, 0x0d, 0x33, 0x27, 0xe8, 0x6f, 0x34, 0xb2, 0x12, 0x73,
-	0xe1, 0xd8, 0xa1, 0xe5, 0x85, 0x09, 0x8f, 0x1f, 0xda, 0xbe, 0x25, 0xf4, 0xe9, 0x8a, 0xb6, 0x75,
-	0xbd, 0xd6, 0xea, 0x49, 0xb6, 0x94, 0x92, 0xf7, 0x32, 0xae, 0xd1, 0x97, 0xec, 0x25, 0xf4, 0x34,
-	0x84, 0x0f, 0xa7, 0xe8, 0xf5, 0x37, 0xef, 0xde, 0x35, 0x9e, 0x48, 0x36, 0xe5, 0x85, 0x49, 0xef,
-	0xa2, 0xba, 0x36, 0xce, 0xfc, 0xc9, 0x45, 0xf5, 0x1a, 0xd8, 0x99, 0xc3, 0x41, 0xe8, 0xdf, 0x35,
-	0x42, 0x9b, 0xc2, 0x3a, 0xb5, 0x13, 0xa7, 0xcd, 0x63, 0x8b, 0x87, 0xf6, 0x91, 0xcf, 0x5d, 0x7d,
-	0xa6, 0xa2, 0x6d, 0xcd, 0xd4, 0x7e, 0xa5, 0x5d, 0x49, 0xb6, 0x5c, 0x6f, 0x7c, 0x90, 0xb2, 0xef,
-	0xa5, 0x64, 0x4f, 0xb2, 0xe5, 0xa6, 0x28, 0x63, 0x7d, 0xc9, 0x5e, 0x4e, 0x37, 0xc1, 0x10, 0x31,
-	0xac, 0x36, 0xdf, 0xe3, 0xeb, 0x63, 0x0d, 0x41, 0x27, 0x58, 0x9c, 0x5f, 0x56, 0x47, 0xc2, 0x9a,
-	0x23, 0x41, 0xe9, 0xdf, 0xca, 0xe2, 0x5d, 0xee, 0xdb, 0x5d, 0x4b, 0xe8, 0xb3, 0x15, 0x6d, 0x4b,
-	0xab, 0x7d, 0x06, 0xe2, 0x97, 0x0a, 0x2f, 0x7b, 0x40, 0x36, 0x20, 0xcf, 0x85, 0x9b, 0x14, 0xea,
-	0x4b, 0xf6, 0x62, 0x59, 0x7a, 0x8a, 0x0f, 0x2b, 0x7f, 0xed, 0x2e, 0xe8, 0x5e, 0x1b, 0x67, 0xf5,
-	0xe4, 0xa2, 0x3a, 0xf9, 0xda, 0xdd, 0xf3, 0xcb, 0xea, 0x70, 0x38, 0x73, 0x38, 0x18, 0xfd, 0x29,
-	0x99, 0xf7, 0x5a, 0x61, 0x14, 0x73, 0xab, 0xc3, 0xe3, 0x40, 0xe8, 0x04, 0x13, 0xfd, 0x4e, 0x4f,
-	0xb2, 0xb9, 0x14, 0x3f, 0x00, 0xb8, 0x2f, 0xd9, 0xcd, 0xb4, 0x4c, 0x0c, 0xb0, 0x62, 0xdf, 0x2e,
-	0x0f, 0x83, 0xa6, 0x3a, 0x95, 0xfe, 0x5c, 0x23, 0x8b, 0xf6, 0x49, 0x12, 0x59, 0x61, 0x14, 0x07,
-	0xb6, 0xef, 0x3d, 0xe2, 0xfa, 0x1c, 0x06, 0xf9, 0xb0, 0x27, 0xd9, 0x02, 0x30, 0xef, 0xe7, 0x44,
-	0xf1, 0xe9, 0x25, 0xf4, 0x9b, 0x96, 0x8c, 0x8e, 0x5a, 0xe5, 0xeb, 0x65, 0x96, 0xfd, 0xd2, 0x88,
-	0x2c, 0x04, 0x5e, 0x68, 0xb9, 0x9e, 0x38, 0xb6, 0x9a, 0x31, 0xe7, 0xfa, 0x7c, 0x45, 0xdb, 0x9a,
-	0xdb, 0x99, 0xcf, 0xcf, 0x53, 0xc3, 0x7b, 0xc4, 0x6b, 0xef, 0x64, 0x47, 0x67, 0x2e, 0xf0, 0xc2,
-	0x3d, 0x4f, 0x1c, 0xd7, 0x63, 0x0e, 0x8a, 0x18, 0x2a, 0x52, 0x30, 0x75, 0x0d, 0x2a, 0xb7, 0x8d,
-	0x27, 0x17, 0xd5, 0xa9, 0xd7, 0x2a, 0xb7, 0x4d, 0x75, 0x1a, 0x6d, 0x11, 0x32, 0xb8, 0xe0, 0xf5,
-	0x05, 0x8c, 0xc6, 0xf2, 0x68, 0x3f, 0x2c, 0x98, 0xf2, 0xd9, 0x7d, 0x21, 0x13, 0xa0, 0x4c, 0xed,
-	0x4b, 0xb6, 0x8c, 0xf1, 0x07, 0x90, 0x61, 0x2a, 0x3c, 0x7d, 0x87, 0x4c, 0x3b, 0x51, 0xc7, 0xe3,
-	0xb1, 0xd0, 0x17, 0xf1, 0xe8, 0x3e, 0x0f, 0x87, 0x3f, 0x83, 0x8a, 0xfb, 0x35, 0x1b, 0xe7, 0xc7,
-	0xd2, 0xcc, 0x0d, 0xe8, 0x3f, 0x34, 0x72, 0x13, 0x5a, 0x0b, 0x1e, 0x5b, 0x81, 0x7d, 0x66, 0x75,
-	0x78, 0xe8, 0x7a, 0x61, 0xcb, 0x3a, 0xf6, 0x8e, 0xf4, 0x25, 0x74, 0xf7, 0x5b, 0xd8, 0xb5, 0xab,
-	0x07, 0x68, 0xb2, 0x6f, 0x9f, 0x1d, 0xa4, 0x06, 0xf7, 0xbd, 0x5a, 0x4f, 0xb2, 0xd5, 0xce, 0x28,
-	0xdc, 0x97, 0xec, 0xa9, 0xb4, 0x7a, 0x8e, 0x72, 0x4a, 0x55, 0x18, 0x3b, 0x75, 0x3c, 0x7c, 0x7e,
-	0x59, 0x1d, 0x17, 0xdf, 0x1c, 0x63, 0x7b, 0x04, 0xe9, 0x68, 0xdb, 0xa2, 0x0d, 0xe9, 0x58, 0x1e,
-	0xa4, 0x23, 0x83, 0x8a, 0x74, 0x64, 0xe3, 0x41, 0x3a, 0x32, 0x80, 0xbe, 0x4b, 0xae, 0x63, 0x93,
-	0xa5, 0xaf, 0x60, 0x11, 0x5f, 0xc9, 0x57, 0x0c, 0xe2, 0x3f, 0x00, 0xa2, 0xa6, 0xc3, 0x2d, 0x87,
-	0x36, 0x7d, 0xc9, 0xe6, 0xd0, 0x1b, 0x8e, 0x0c, 0x33, 0x45, 0xe9, 0x7d, 0xb2, 0x90, 0x1d, 0x28,
-	0x97, 0xfb, 0x3c, 0xe1, 0x3a, 0xc5, 0xcd, 0xfe, 0x02, 0xb6, 0x14, 0x48, 0xec, 0x21, 0xde, 0x97,
-	0x8c, 0x2a, 0x47, 0x2a, 0x05, 0x0d, 0xb3, 0x64, 0x43, 0xcf, 0x88, 0x8e, 0x05, 0xba, 0x13, 0x47,
-	0xad, 0x98, 0x0b, 0xa1, 0x56, 0xea, 0x55, 0xfc, 0x3e, 0xb8, 0x75, 0xd7, 0xc1, 0xe6, 0x20, 0x33,
-	0x51, 0xeb, 0x75, 0x7a, 0x8f, 0x8d, 0x65, 0x8b, 0x6f, 0x1f, 0x3f, 0x99, 0x36, 0xc8, 0x62, 0xb6,
-	0x2f, 0x3a, 0xf6, 0x89, 0xe0, 0x96, 0xd0, 0xd7, 0x30, 0xde, 0xab, 0xf0, 0x1d, 0x29, 0x73, 0x00,
-	0x44, 0xa3, 0xf8, 0x0e, 0x15, 0x2c, 0xbc, 0x97, 0x4c, 0x29, 0x27, 0x0b, 0xb0, 0xcb, 0x20, 0xa9,
-	0xbe, 0xe7, 0x24, 0x42, 0x5f, 0x47, 0x9f, 0xdf, 0x01, 0x9f, 0x81, 0x7d, 0xb6, 0x9b, 0xe3, 0x83,
-	0x53, 0xa7, 0x80, 0xe5, 0xd2, 0x97, 0x05, 0x48, 0x2b, 0x9d, 0x59, 0x9a, 0x4d, 0x5d, 0xb2, 0xe6,
-	0x7a, 0x02, 0x4a, 0xb2, 0x25, 0x3a, 0x76, 0x2c, 0xb8, 0x85, 0x37, 0xbf, 0x7e, 0x13, 0x57, 0x02,
-	0x7b, 0xad, 0x8c, 0x6f, 0x20, 0x8d, 0x3d, 0x45, 0xd1, 0x6b, 0x8d, 0x52, 0x86, 0x39, 0xc6, 0x5e,
-	0x8d, 0x92, 0xf0, 0xa0, 0x63, 0x79, 0xa1, 0xcb, 0xcf, 0xb8, 0xd0, 0x6f, 0x8d, 0x44, 0x39, 0xe4,
-	0x41, 0xe7, 0x5e, 0xca, 0x0e, 0x47, 0x51, 0xa8, 0x41, 0x14, 0x05, 0xa4, 0x3b, 0xe4, 0x06, 0x2e,
-	0x80, 0xab, 0xeb, 0xe8, 0x77, 0xa3, 0x27, 0x59, 0x86, 0x14, 0x57, 0x7b, 0x3a, 0x34, 0xcc, 0x0c,
-	0xa7, 0x09, 0xb9, 0x75, 0xca, 0xed, 0x63, 0x0b, 0x76, 0xb5, 0x95, 0xb4, 0x63, 0x2e, 0xda, 0x91,
-	0xef, 0x5a, 0x1d, 0x27, 0xd1, 0x9f, 0xc2, 0x84, 0x43, 0x79, 0x5f, 0x03, 0x93, 0xef, 0xda, 0xa2,
-	0x7d, 0x98, 0x1b, 0x1c, 0x38, 0x49, 0x5f, 0xb2, 0x0d, 0x74, 0x39, 0x8e, 0x2c, 0x16, 0x75, 0xec,
-	0x54, 0xba, 0x4b, 0xe6, 0x02, 0x3b, 0x3e, 0xe6, 0xb1, 0x15, 0xda, 0x01, 0xd7, 0x37, 0xb0, 0xab,
-	0x32, 0xa0, 0x9c, 0xa5, 0xf0, 0xfb, 0x76, 0xc0, 0x8b, 0x72, 0x36, 0x80, 0x0c, 0x53, 0xe1, 0x69,
-	0x97, 0x6c, 0xc0, 0xeb, 0xc5, 0x8a, 0x4e, 0x43, 0x1e, 0x8b, 0xb6, 0xd7, 0xb1, 0x9a, 0x71, 0x14,
-	0x58, 0x1d, 0x3b, 0xe6, 0x61, 0xa2, 0x3f, 0x8d, 0x29, 0xf8, 0xff, 0x9e, 0x64, 0xb7, 0xc0, 0xea,
-	0x41, 0x6e, 0x54, 0x8f, 0xa3, 0xe0, 0x00, 0x4d, 0xfa, 0x92, 0x3d, 0x9b, 0x57, 0xbc, 0x71, 0xbc,
-	0x61, 0x7e, 0xd3, 0x4c, 0xfa, 0x0b, 0x8d, 0xac, 0x04, 0x91, 0x6b, 0x25, 0x5e, 0xc0, 0xad, 0x53,
-	0x2f, 0x74, 0xa3, 0x53, 0x4b, 0xe8, 0xcf, 0x60, 0xc2, 0x3e, 0xba, 0x92, 0x6c, 0xc5, 0xb4, 0x4f,
-	0xf7, 0x23, 0xf7, 0xd0, 0x0b, 0xf8, 0x07, 0xc8, 0xc2, 0xe5, 0xbd, 0x18, 0x94, 0x90, 0xa2, 0xf7,
-	0x2c, 0xc3, 0x79, 0xe6, 0xce, 0x2f, 0xab, 0xa3, 0x5e, 0xcc, 0x21, 0x1f, 0xf4, 0x53, 0x8d, 0xac,
-	0x67, 0xc7, 0xc4, 0x39, 0x89, 0x41, 0x9b, 0x75, 0x1a, 0x7b, 0x09, 0x17, 0xfa, 0xb3, 0x28, 0xe6,
-	0xfb, 0x50, 0x7a, 0xd3, 0x0d, 0x9f, 0xf1, 0x1f, 0x20, 0xdd, 0x97, 0xec, 0xb6, 0x72, 0x6a, 0x4a,
-	0x9c, 0x72, 0x78, 0x76, 0x94, 0xb3, 0xa3, 0xed, 0x98, 0xe3, 0x3c, 0x41, 0x11, 0xcb, 0xf7, 0x76,
-	0x13, 0x9e, 0x4a, 0xfa, 0xe6, 0xa0, 0x88, 0x65, 0x44, 0x1d, 0xf0, 0xe2, 0xf0, 0xab, 0xa0, 0x61,
-	0x96, 0x6c, 0xa8, 0x4f, 0x96, 0xf1, 0x09, 0x6b, 0x41, 0x2d, 0xb0, 0xd2, 0xfa, 0xca, 0xb0, 0xbe,
-	0xde, 0xcc, 0xeb, 0x6b, 0x0d, 0xf8, 0x41, 0x91, 0xc5, 0xae, 0xfe, 0xa8, 0x84, 0x15, 0x99, 0x2d,
-	0xc3, 0x86, 0x39, 0x64, 0x47, 0x3f, 0xd7, 0xc8, 0x0a, 0x6e, 0x21, 0x7c, 0x01, 0x5b, 0xe9, 0x13,
-	0x58, 0xaf, 0x60, 0xbc, 0x55, 0x78, 0x41, 0xec, 0x46, 0x9d, 0xae, 0x09, 0xdc, 0x3e, 0x52, 0xb5,
-	0xfb, 0xd0, 0x83, 0x39, 0x65, 0xb0, 0x2f, 0xd9, 0x56, 0xb1, 0x8d, 0x14, 0x5c, 0x49, 0xa3, 0x48,
-	0xec, 0xd0, 0xb5, 0x63, 0x17, 0xee, 0xff, 0x99, 0x7c, 0x60, 0x0e, 0x3b, 0xa2, 0x7f, 0x04, 0x39,
-	0x36, 0x14, 0x50, 0x1e, 0x0a, 0x2f, 0xf1, 0x1e, 0x42, 0x46, 0xf5, 0xe7, 0x30, 0x9d, 0x67, 0xd0,
-	0x10, 0xee, 0xda, 0x82, 0x37, 0x72, 0xae, 0x8e, 0x0d, 0xa1, 0x53, 0x86, 0xfa, 0x92, 0xad, 0xa7,
-	0x62, 0xca, 0x38, 0xf4, 0x40, 0x23, 0xb6, 0xa3, 0x10, 0xb4, 0x81, 0x43, 0x41, 0xcc, 0x21, 0x1b,
-	0x41, 0xff, 0xa0, 0x91, 0xe5, 0x66, 0xe4, 0xfb, 0xd1, 0xa9, 0xf5, 0xf1, 0x49, 0xe8, 0x40, 0x3b,
-	0x22, 0x74, 0x63, 0xa0, 0xf2, 0x7b, 0x39, 0xf8, 0xae, 0xd8, 0xf3, 0x62, 0x01, 0x2a, 0x3f, 0x2e,
-	0x43, 0x85, 0xca, 0x21, 0x1c, 0x55, 0x0e, 0xdb, 0x8e, 0x42, 0xa0, 0x72, 0x28, 0x88, 0xb9, 0x94,
-	0x2a, 0x2a, 0x60, 0xfa, 0x80, 0x2c, 0xc2, 0x8e, 0x1a, 0x54, 0x07, 0xfd, 0x79, 0x94, 0x08, 0x0f,
-	0xab, 0x05, 0x60, 0x8a, 0x73, 0xdd, 0x97, 0x6c, 0x35, 0xbd, 0xfc, 0x54, 0xd4, 0x30, 0xcb, 0x56,
-	0xe8, 0x90, 0x87, 0xae, 0xe2, 0xb0, 0xaa, 0x38, 0xe4, 0xa1, 0x3b, 0xc6, 0xa1, 0x8a, 0x82, 0x43,
-	0x75, 0x0c, 0x45, 0x10, 0x15, 0x9e, 0x41, 0x37, 0x2a, 0xf4, 0xdb, 0xe8, 0x0d, 0x8b, 0x20, 0xc0,
-	0x3f, 0x42, 0xb4, 0x28, 0x82, 0x03, 0xc8, 0x30, 0x15, 0x1e, 0x9d, 0x80, 0xaa, 0xcc, 0xc9, 0x0b,
-	0x8a, 0x13, 0x1e, 0xba, 0xc3, 0x4e, 0x0a, 0x08, 0x9c, 0x14, 0x03, 0x68, 0xec, 0x71, 0x3e, 0xdc,
-	0x7d, 0x09, 0x8f, 0xf5, 0x17, 0xb1, 0x07, 0x5d, 0xcd, 0x4f, 0x1c, 0x5a, 0xd5, 0x91, 0xaa, 0x6d,
-	0xe5, 0x8d, 0xef, 0xd9, 0x00, 0xec, 0x4b, 0xb6, 0x82, 0xfe, 0x15, 0xcc, 0x30, 0x55, 0x0b, 0x7a,
-	0x4c, 0x66, 0x63, 0x6e, 0xbb, 0x56, 0x14, 0xfa, 0x5d, 0xfd, 0xcf, 0x75, 0x54, 0xb9, 0x7f, 0x25,
-	0x19, 0xdd, 0xe3, 0x9d, 0x98, 0x3b, 0x76, 0xc2, 0x5d, 0x93, 0xdb, 0xee, 0x83, 0xd0, 0xef, 0xf6,
-	0x24, 0xd3, 0x5e, 0x2d, 0xfe, 0x44, 0x89, 0x23, 0x6c, 0xd6, 0x5f, 0x89, 0x02, 0x0f, 0x6e, 0xce,
-	0xa4, 0x8b, 0x7f, 0xa2, 0x8c, 0xa0, 0xba, 0x66, 0xce, 0xc4, 0x99, 0x03, 0xfa, 0x33, 0xb2, 0x52,
-	0xea, 0xe0, 0xf1, 0x36, 0xfb, 0x4b, 0x1d, 0x5f, 0x56, 0xef, 0x5d, 0x49, 0xa6, 0x0f, 0x82, 0xee,
-	0x0f, 0xfa, 0xf0, 0x03, 0x27, 0xc9, 0x43, 0x6f, 0x0e, 0xb7, 0xf1, 0x07, 0x4e, 0xa2, 0x28, 0xd0,
-	0x35, 0x73, 0xb1, 0x4c, 0xd2, 0x1f, 0x93, 0xe9, 0xb4, 0x7b, 0x11, 0xfa, 0x97, 0x75, 0xac, 0xbc,
-	0xdf, 0x82, 0x6b, 0x60, 0x10, 0x28, 0xed, 0x4a, 0x45, 0xf9, 0xe3, 0xb2, 0x29, 0x8a, 0xeb, 0xac,
-	0xdc, 0xea, 0x9a, 0x99, 0xfb, 0xa3, 0xc7, 0x64, 0x11, 0xfb, 0xba, 0xc1, 0xbe, 0xfb, 0x6b, 0x9a,
-	0xbf, 0xdd, 0x2b, 0xc9, 0x6e, 0x0d, 0x22, 0x34, 0x1c, 0x3b, 0x2c, 0x36, 0x57, 0x1e, 0xe7, 0xd9,
-	0xa2, 0xab, 0x2b, 0xa8, 0xf2, 0x87, 0x2c, 0x94, 0x38, 0xe3, 0xb3, 0x29, 0x32, 0xa7, 0x2c, 0x37,
-	0xfd, 0x88, 0x4c, 0xf3, 0x30, 0x89, 0x3d, 0x2e, 0x74, 0x0d, 0xff, 0x56, 0xd0, 0xc7, 0x6c, 0x8a,
-	0xf7, 0xc2, 0x24, 0xee, 0xd6, 0x5e, 0xcc, 0xff, 0x4d, 0xc8, 0x26, 0x14, 0x3d, 0x2f, 0x8c, 0x71,
-	0xd9, 0xae, 0xe3, 0x2f, 0x33, 0x37, 0xa0, 0xbf, 0xcb, 0x2e, 0x2f, 0xe1, 0x85, 0x2d, 0x9f, 0x5b,
-	0xc8, 0x5a, 0x02, 0x1e, 0x7d, 0x93, 0x98, 0xc2, 0x26, 0xf4, 0x45, 0x81, 0x7d, 0xd6, 0x40, 0x1e,
-	0xa3, 0x34, 0xd4, 0x97, 0xdf, 0x28, 0x55, 0xea, 0xfb, 0x76, 0xde, 0x50, 0x1e, 0x11, 0x63, 0xfc,
-	0xc0, 0x03, 0x10, 0xac, 0xcc, 0x31, 0x1c, 0x7d, 0x44, 0x16, 0x41, 0x5a, 0x12, 0x25, 0xd0, 0x40,
-	0x83, 0xa6, 0x29, 0xd4, 0x74, 0x98, 0xf5, 0x9f, 0x87, 0x40, 0x64, 0x6a, 0x9e, 0xcb, 0xd5, 0x14,
-	0xa0, 0xa2, 0xe3, 0x8d, 0xbb, 0x6f, 0xbd, 0xa9, 0xe8, 0x28, 0xcd, 0x05, 0x05, 0xc0, 0x9b, 0x25,
-	0xd4, 0xf8, 0xbd, 0x46, 0x96, 0x87, 0xd3, 0x0b, 0xcf, 0x8d, 0x00, 0x5e, 0xe3, 0xd9, 0x3f, 0x73,
-	0xff, 0x07, 0x6f, 0x0b, 0x04, 0x94, 0x3e, 0x29, 0x71, 0xda, 0xc5, 0x4b, 0x9b, 0x0c, 0x86, 0x66,
-	0x6a, 0x48, 0xeb, 0xe4, 0x06, 0x3c, 0xdc, 0xbd, 0x04, 0xf3, 0x3b, 0x53, 0xdb, 0xc6, 0xfe, 0x10,
-	0x91, 0xe2, 0x08, 0xa7, 0xc3, 0xc2, 0xcb, 0x9c, 0x32, 0x36, 0x33, 0xdb, 0xda, 0xfd, 0xaf, 0xbe,
-	0xde, 0x9c, 0xb8, 0xfc, 0x7a, 0x73, 0xe2, 0xab, 0xab, 0x4d, 0xed, 0xf2, 0x6a, 0x53, 0xfb, 0xf5,
-	0xe3, 0xcd, 0x89, 0x2f, 0x1e, 0x6f, 0x6a, 0x97, 0x8f, 0x37, 0x27, 0xfe, 0xf5, 0x78, 0x73, 0xe2,
-	0xc3, 0x97, 0xfe, 0x8b, 0x3f, 0x52, 0xd3, 0x7d, 0x74, 0x74, 0x03, 0xff, 0x50, 0x7d, 0xfd, 0x3f,
-	0x01, 0x00, 0x00, 0xff, 0xff, 0xe3, 0x79, 0xd7, 0x43, 0x6e, 0x17, 0x00, 0x00,
+	// 2460 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xa4, 0x58, 0x4d, 0x6c, 0x24, 0x47,
+	0x15, 0x76, 0xdb, 0xfb, 0x63, 0x97, 0xd7, 0x5e, 0xbb, 0x6c, 0xef, 0x76, 0x9c, 0xc4, 0x35, 0xe9,
+	0xcc, 0x26, 0x93, 0x90, 0x78, 0x37, 0x4e, 0x14, 0x29, 0x11, 0x01, 0x32, 0x76, 0x46, 0x2c, 0x8b,
+	0xb3, 0x56, 0x8f, 0x21, 0x90, 0x20, 0x35, 0xed, 0xee, 0x9a, 0x99, 0x8e, 0x7b, 0xba, 0x87, 0xae,
+	0xf6, 0xda, 0xb3, 0x87, 0x28, 0xe4, 0x80, 0x90, 0xc8, 0x01, 0x99, 0x03, 0xe2, 0x80, 0x14, 0x09,
+	0x84, 0x20, 0x5c, 0x38, 0x73, 0xe1, 0xba, 0x17, 0x64, 0x9f, 0x10, 0xe2, 0x50, 0x52, 0xbc, 0xb7,
+	0x39, 0xf6, 0x71, 0x4f, 0xe8, 0xbd, 0xfe, 0x99, 0xea, 0x99, 0x89, 0x84, 0xc4, 0x6d, 0xea, 0xfb,
+	0x5e, 0xbd, 0xf7, 0x75, 0xd5, 0xab, 0x57, 0xaf, 0x86, 0x54, 0x7d, 0xef, 0xe0, 0xb6, 0x13, 0x06,
+	0x2d, 0xaf, 0x7d, 0xbb, 0x15, 0xfa, 0x2e, 0x8f, 0xd2, 0xc1, 0x51, 0x64, 0xc7, 0x5e, 0x18, 0x6c,
+	0xf6, 0xa2, 0x30, 0x0e, 0xe9, 0x95, 0x14, 0x5c, 0x7f, 0x7a, 0xcc, 0x3a, 0xee, 0xf7, 0x78, 0x6a,
+	0xb4, 0xbe, 0xa6, 0x90, 0xc2, 0x7b, 0x98, 0xc3, 0xeb, 0x0a, 0xdc, 0x3b, 0xf2, 0xfd, 0x30, 0x72,
+	0x79, 0x94, 0x71, 0x35, 0x85, 0x7b, 0xc0, 0x23, 0xe1, 0x85, 0x81, 0x17, 0xb4, 0x27, 0x28, 0x58,
+	0x67, 0x8a, 0xe5, 0x81, 0x1f, 0x3a, 0x87, 0xa3, 0xae, 0x28, 0x18, 0xb4, 0xc4, 0x6d, 0x10, 0x24,
+	0x32, 0xec, 0x99, 0x0c, 0x73, 0xc2, 0x5e, 0x3f, 0xb2, 0x83, 0x36, 0xef, 0xf2, 0xb8, 0x13, 0xba,
+	0x19, 0x3b, 0xc7, 0x4f, 0xe2, 0xf4, 0xa7, 0xf1, 0xaf, 0x19, 0xf2, 0x54, 0x03, 0xbf, 0x67, 0x87,
+	0x3f, 0xf0, 0x1c, 0xbe, 0xad, 0x2a, 0xa0, 0x5f, 0x6a, 0x64, 0xce, 0x45, 0xdc, 0xf2, 0x5c, 0x5d,
+	0xab, 0x68, 0xb5, 0x6b, 0xf5, 0xcf, 0xb5, 0x47, 0x92, 0x4d, 0xfd, 0x47, 0xb2, 0x37, 0xda, 0x5e,
+	0xdc, 0x39, 0x3a, 0xd8, 0x74, 0xc2, 0xee, 0x6d, 0xd1, 0x0f, 0x9c, 0xb8, 0xe3, 0x05, 0x6d, 0xe5,
+	0x17, 0x48, 0xc0, 0x20, 0x4e, 0xe8, 0x6f, 0xa6, 0xde, 0xef, 0xee, 0x5c, 0x48, 0x36, 0x9b, 0xff,
+	0x1e, 0x48, 0x36, 0xeb, 0x66, 0xbf, 0x13, 0xc9, 0x16, 0x4e, 0xba, 0xfe, 0xdb, 0x86, 0xe7, 0xbe,
+	0x62, 0xc7, 0x71, 0x64, 0x0c, 0xce, 0xaa, 0x57, 0xb3, 0xdf, 0xc9, 0x59, 0xb5, 0xb0, 0xfb, 0xe5,
+	0x79, 0x55, 0x3b, 0x3d, 0xaf, 0x16, 0x3e, 0xcc, 0x9c, 0x71, 0xe9, 0x9f, 0x34, 0xb2, 0xe0, 0x05,
+	0x71, 0x14, 0xba, 0x47, 0x0e, 0x77, 0xad, 0x83, 0xbe, 0x3e, 0x8d, 0x82, 0x3f, 0xfd, 0xbf, 0x04,
+	0x0f, 0x24, 0xbb, 0x36, 0xf4, 0x5a, 0xef, 0x27, 0x92, 0xdd, 0x4c, 0x85, 0x2a, 0x60, 0x21, 0x79,
+	0x79, 0x0c, 0x05, 0xc1, 0x66, 0xc9, 0x03, 0x75, 0xc8, 0x0a, 0x0f, 0x9c, 0xa8, 0xdf, 0x83, 0x35,
+	0xb6, 0x7a, 0xb6, 0x10, 0xc7, 0x61, 0xe4, 0xea, 0x33, 0x15, 0xad, 0x36, 0x57, 0xdf, 0x1a, 0x48,
+	0x46, 0x87, 0xf4, 0x5e, 0xc6, 0x26, 0x92, 0xe9, 0x18, 0x76, 0x9c, 0x32, 0xcc, 0x09, 0xf6, 0xc6,
+	0x3f, 0x6e, 0x91, 0x95, 0x74, 0x63, 0xcb, 0x5b, 0xda, 0x24, 0xd3, 0xd9, 0x56, 0xce, 0xd5, 0xb7,
+	0x2f, 0x24, 0x9b, 0xc6, 0x4f, 0x9c, 0xf6, 0x20, 0xc2, 0x46, 0x69, 0x07, 0x2a, 0x41, 0xe8, 0xf2,
+	0x96, 0x7d, 0xe4, 0xc7, 0x6f, 0x1b, 0x71, 0x74, 0xc4, 0xd5, 0x2d, 0x39, 0x3d, 0xaf, 0x4e, 0xdf,
+	0xdd, 0xf9, 0x02, 0xbe, 0x6d, 0xda, 0x73, 0xe9, 0x0f, 0xc8, 0x65, 0xdf, 0x3e, 0xe0, 0x3e, 0xae,
+	0xf8, 0x5c, 0xfd, 0xdb, 0x03, 0xc9, 0x52, 0x20, 0x91, 0xac, 0x82, 0x4e, 0x71, 0x94, 0xf9, 0x8d,
+	0xb8, 0x88, 0xed, 0x28, 0x7e, 0xdb, 0x68, 0xd9, 0xbe, 0x40, 0xb7, 0x64, 0x48, 0x7f, 0x7a, 0x5e,
+	0x9d, 0x32, 0xd3, 0xc9, 0xb4, 0x4d, 0xae, 0xb7, 0x3c, 0x9f, 0x8b, 0xbe, 0x88, 0x79, 0xd7, 0x82,
+	0xfc, 0xc6, 0x45, 0x5a, 0xdc, 0xa2, 0x9b, 0x2d, 0xb1, 0xd9, 0x28, 0xa8, 0xfd, 0x7e, 0x8f, 0xd7,
+	0x5f, 0x1e, 0x48, 0xb6, 0xd8, 0x2a, 0x61, 0x89, 0x64, 0xab, 0x18, 0xbd, 0x0c, 0x1b, 0xe6, 0x88,
+	0x1d, 0xdd, 0x25, 0x97, 0x7a, 0x76, 0xdc, 0xd1, 0x2f, 0xa1, 0xfc, 0xb7, 0x06, 0x92, 0xe1, 0x38,
+	0x91, 0xec, 0x69, 0x9c, 0x0f, 0x83, 0x4c, 0x7c, 0xb1, 0x24, 0x9f, 0x80, 0xf0, 0xb9, 0x82, 0x79,
+	0x72, 0x56, 0xd5, 0x3e, 0x31, 0x71, 0x1a, 0xdd, 0x23, 0x97, 0x50, 0xec, 0xe5, 0x4c, 0x6c, 0x7a,
+	0x7a, 0x37, 0xd3, 0xed, 0x40, 0xb1, 0x35, 0x08, 0x11, 0xa7, 0x12, 0xaf, 0x63, 0x08, 0x18, 0x14,
+	0x69, 0x34, 0x57, 0x8c, 0x4c, 0xb4, 0xa2, 0x3f, 0x21, 0x57, 0xd3, 0x3c, 0x17, 0xfa, 0x95, 0xca,
+	0x4c, 0x6d, 0x7e, 0xeb, 0xb9, 0xb2, 0xd3, 0x09, 0x87, 0xb7, 0xce, 0x20, 0xed, 0x07, 0x92, 0xe5,
+	0x33, 0x13, 0xc9, 0xae, 0x61, 0xa8, 0x74, 0x6c, 0x98, 0x39, 0x41, 0x7f, 0xa3, 0x91, 0xe5, 0x88,
+	0x0b, 0xc7, 0x0e, 0x2c, 0x2f, 0x88, 0x79, 0xf4, 0xc0, 0xf6, 0x2d, 0xa1, 0x5f, 0xad, 0x68, 0xb5,
+	0xcb, 0xf5, 0xf6, 0x40, 0xb2, 0xeb, 0x29, 0x79, 0x37, 0xe3, 0x9a, 0x89, 0x64, 0x2f, 0xa1, 0xa7,
+	0x11, 0x7c, 0x74, 0x89, 0x5e, 0x7f, 0xf3, 0xce, 0x1d, 0xe3, 0x89, 0x64, 0x33, 0x5e, 0x10, 0x0f,
+	0xce, 0xaa, 0xab, 0x93, 0xcc, 0x9f, 0x9c, 0x55, 0x2f, 0x81, 0x9d, 0x39, 0x1a, 0x84, 0xfe, 0x5d,
+	0x23, 0xb4, 0x25, 0xac, 0x63, 0x3b, 0x76, 0x3a, 0x3c, 0xb2, 0x78, 0x60, 0x1f, 0xf8, 0xdc, 0xd5,
+	0x67, 0x2b, 0x5a, 0x6d, 0xb6, 0xfe, 0x2b, 0xed, 0x42, 0xb2, 0xa5, 0x46, 0xf3, 0x83, 0x94, 0x7d,
+	0x2f, 0x25, 0x07, 0x92, 0x2d, 0xb5, 0x44, 0x19, 0x4b, 0x24, 0x7b, 0x39, 0x4d, 0x82, 0x11, 0x62,
+	0x54, 0x6d, 0x9e, 0xe3, 0x6b, 0x13, 0x0d, 0x41, 0x27, 0x58, 0x9c, 0x9e, 0x57, 0xc7, 0xc2, 0x9a,
+	0x63, 0x41, 0xe9, 0xdf, 0xca, 0xe2, 0x5d, 0xee, 0xdb, 0x7d, 0x4b, 0xe8, 0x73, 0x15, 0xad, 0xa6,
+	0xd5, 0x3f, 0x03, 0xf1, 0xd7, 0x0b, 0x2f, 0x3b, 0x40, 0x36, 0x61, 0x9d, 0x0b, 0x37, 0x29, 0x94,
+	0x48, 0xf6, 0x62, 0x59, 0x7a, 0x8a, 0x8f, 0x2a, 0x7f, 0xed, 0x0e, 0xe8, 0x5e, 0x9d, 0x64, 0xf5,
+	0xe4, 0xac, 0x3a, 0xfd, 0xda, 0x9d, 0xd3, 0xf3, 0xea, 0x68, 0x38, 0x73, 0x34, 0x18, 0x14, 0xfb,
+	0x55, 0x45, 0x72, 0xec, 0x75, 0x79, 0x78, 0x14, 0x5b, 0x42, 0xaf, 0xa1, 0xe8, 0xfe, 0x85, 0x64,
+	0xcb, 0x85, 0x93, 0xfd, 0x94, 0x05, 0xd5, 0xcb, 0x85, 0xa3, 0x1c, 0x4c, 0x24, 0x7b, 0xa6, 0xac,
+	0x3b, 0x67, 0x8a, 0x0c, 0xbf, 0x31, 0x99, 0x3a, 0x3d, 0xaf, 0x8e, 0xc7, 0x30, 0xc7, 0x23, 0xd0,
+	0x9f, 0x92, 0x6b, 0x5e, 0x3b, 0x08, 0x23, 0x6e, 0xf5, 0x78, 0xd4, 0x15, 0x3a, 0xc1, 0xac, 0x78,
+	0x67, 0x20, 0xd9, 0x7c, 0x8a, 0xef, 0x01, 0x9c, 0x48, 0x76, 0x23, 0xad, 0x69, 0x43, 0xac, 0x90,
+	0xb0, 0x34, 0x0a, 0x9a, 0xea, 0x54, 0xfa, 0x73, 0x8d, 0x2c, 0xda, 0x47, 0x71, 0x68, 0x05, 0x61,
+	0xd4, 0xb5, 0x7d, 0xef, 0x21, 0xd7, 0xe7, 0x31, 0xc8, 0x87, 0x03, 0xc9, 0x16, 0x80, 0x79, 0x3f,
+	0x27, 0x8a, 0x7d, 0x2a, 0xa1, 0x5f, 0x97, 0x5f, 0x74, 0xdc, 0x2a, 0x4f, 0x2e, 0xb3, 0xec, 0x97,
+	0x86, 0x64, 0xa1, 0xeb, 0x05, 0x96, 0xeb, 0x89, 0x43, 0xab, 0x15, 0x71, 0xae, 0x5f, 0xab, 0x68,
+	0xb5, 0xf9, 0xad, 0x6b, 0xf9, 0xe1, 0x6f, 0x7a, 0x0f, 0x79, 0xfd, 0x9d, 0xec, 0x9c, 0xcf, 0x77,
+	0xbd, 0x60, 0xc7, 0x13, 0x87, 0x8d, 0x88, 0x83, 0x22, 0x86, 0x8a, 0x14, 0x4c, 0x4d, 0x98, 0xca,
+	0x2d, 0xe3, 0xc9, 0x59, 0x75, 0xe6, 0xb5, 0xca, 0x2d, 0x53, 0x9d, 0x46, 0xdb, 0x84, 0x0c, 0xbb,
+	0x11, 0x7d, 0x01, 0xa3, 0xb1, 0x3c, 0xda, 0x0f, 0x0b, 0xa6, 0x5c, 0x68, 0x5e, 0xc8, 0x04, 0x28,
+	0x53, 0x13, 0xc9, 0x96, 0x30, 0xfe, 0x10, 0x32, 0x4c, 0x85, 0xa7, 0xef, 0x90, 0xab, 0x4e, 0xd8,
+	0xf3, 0x78, 0x24, 0xf4, 0x45, 0xac, 0x33, 0xcf, 0x43, 0xa5, 0xca, 0xa0, 0xa2, 0x19, 0xc8, 0xc6,
+	0x79, 0x0d, 0x31, 0x73, 0x03, 0xfa, 0x4f, 0x8d, 0xdc, 0x80, 0x3e, 0x88, 0x47, 0x56, 0xd7, 0x3e,
+	0xb1, 0x7a, 0x3c, 0x70, 0xbd, 0xa0, 0x6d, 0x1d, 0x7a, 0x07, 0xfa, 0x75, 0x74, 0xf7, 0x5b, 0x38,
+	0x62, 0x2b, 0x7b, 0x68, 0xb2, 0x6b, 0x9f, 0xec, 0xa5, 0x06, 0xf7, 0xbc, 0xfa, 0x40, 0xb2, 0x95,
+	0xde, 0x38, 0x9c, 0x48, 0xf6, 0x54, 0x5a, 0xea, 0xc7, 0x39, 0xa5, 0x84, 0x4d, 0x9c, 0x3a, 0x19,
+	0x3e, 0x3d, 0xaf, 0x4e, 0x8a, 0x6f, 0x4e, 0xb0, 0x3d, 0x80, 0xe5, 0xe8, 0xd8, 0xa2, 0x03, 0xcb,
+	0xb1, 0x34, 0x5c, 0x8e, 0x0c, 0x2a, 0x96, 0x23, 0x1b, 0x0f, 0x97, 0x23, 0x03, 0xe8, 0xbb, 0xe4,
+	0x32, 0x76, 0x84, 0xfa, 0x32, 0xde, 0x38, 0xcb, 0xf9, 0x8e, 0x41, 0xfc, 0xfb, 0x40, 0xd4, 0x75,
+	0xb8, 0x92, 0xd1, 0x26, 0x91, 0x6c, 0x1e, 0xbd, 0xe1, 0xc8, 0x30, 0x53, 0x94, 0xde, 0x23, 0x0b,
+	0xd9, 0x81, 0x72, 0xb9, 0xcf, 0x63, 0xae, 0x53, 0x4c, 0xf6, 0x17, 0xb0, 0xff, 0x41, 0x62, 0x07,
+	0xf1, 0x44, 0x32, 0xaa, 0x1c, 0xa9, 0x14, 0x34, 0xcc, 0x92, 0x0d, 0x3d, 0x21, 0x3a, 0xde, 0x26,
+	0xbd, 0x28, 0x6c, 0x47, 0x5c, 0x08, 0xf5, 0x5a, 0x59, 0xc1, 0xef, 0x83, 0x16, 0x61, 0x0d, 0x6c,
+	0xf6, 0x32, 0x13, 0xf5, 0x72, 0x49, 0x2f, 0xdd, 0x89, 0x6c, 0xf1, 0xed, 0x93, 0x27, 0xd3, 0x26,
+	0x59, 0xcc, 0xf2, 0xa2, 0x67, 0x1f, 0x09, 0x6e, 0x09, 0x7d, 0x15, 0xe3, 0xbd, 0x0a, 0xdf, 0x91,
+	0x32, 0x7b, 0x40, 0x34, 0x8b, 0xef, 0x50, 0xc1, 0xc2, 0x7b, 0xc9, 0x94, 0x72, 0xb2, 0x00, 0x59,
+	0x06, 0x8b, 0xea, 0x7b, 0x4e, 0x2c, 0xf4, 0x35, 0xf4, 0xf9, 0x1d, 0xf0, 0xd9, 0xb5, 0x4f, 0xb6,
+	0x73, 0x7c, 0x78, 0xea, 0x14, 0xb0, 0x5c, 0xa7, 0xb3, 0x00, 0x69, 0x59, 0x36, 0x4b, 0xb3, 0xa9,
+	0x4b, 0x56, 0x5d, 0x4f, 0xc0, 0xfd, 0x61, 0x89, 0x9e, 0x1d, 0x09, 0x6e, 0x61, 0x9b, 0xa2, 0xdf,
+	0xc0, 0x9d, 0xc0, 0xc6, 0x30, 0xe3, 0x9b, 0x48, 0x63, 0x03, 0x54, 0x34, 0x86, 0xe3, 0x94, 0x61,
+	0x4e, 0xb0, 0x57, 0xa3, 0xc4, 0xbc, 0xdb, 0xb3, 0xbc, 0xc0, 0xe5, 0x27, 0x5c, 0xe8, 0x37, 0xc7,
+	0xa2, 0xec, 0xf3, 0x6e, 0xef, 0x6e, 0xca, 0x8e, 0x46, 0x51, 0xa8, 0x61, 0x14, 0x05, 0xa4, 0x5b,
+	0xe4, 0x0a, 0x6e, 0x80, 0xab, 0xeb, 0xe8, 0x77, 0x7d, 0x20, 0x59, 0x86, 0x14, 0x7d, 0x48, 0x3a,
+	0x34, 0xcc, 0x0c, 0xa7, 0x31, 0xb9, 0x79, 0xcc, 0xed, 0x43, 0x0b, 0xb2, 0xda, 0x8a, 0x3b, 0x11,
+	0x17, 0x9d, 0xd0, 0x77, 0xad, 0x9e, 0x13, 0xeb, 0x4f, 0xe1, 0x82, 0x43, 0x79, 0x5f, 0x05, 0x93,
+	0xef, 0xda, 0xa2, 0xb3, 0x9f, 0x1b, 0xec, 0x39, 0x71, 0x22, 0xd9, 0x3a, 0xba, 0x9c, 0x44, 0x16,
+	0x9b, 0x3a, 0x71, 0x2a, 0xdd, 0x26, 0xf3, 0x5d, 0x3b, 0x3a, 0xe4, 0x91, 0x15, 0xd8, 0x5d, 0xae,
+	0xaf, 0x63, 0x0b, 0x68, 0x40, 0x39, 0x4b, 0xe1, 0xf7, 0xed, 0x2e, 0x2f, 0xca, 0xd9, 0x10, 0x32,
+	0x4c, 0x85, 0xa7, 0x7d, 0xb2, 0x0e, 0x4f, 0x2d, 0x2b, 0x3c, 0x0e, 0x78, 0x24, 0x3a, 0x5e, 0xcf,
+	0x6a, 0x45, 0x61, 0xd7, 0xea, 0xd9, 0x11, 0x0f, 0x62, 0xfd, 0x69, 0x5c, 0x82, 0x6f, 0x0e, 0x24,
+	0xbb, 0x09, 0x56, 0xf7, 0x73, 0xa3, 0x46, 0x14, 0x76, 0xf7, 0xd0, 0x24, 0x91, 0xec, 0xd9, 0xbc,
+	0xe2, 0x4d, 0xe2, 0x0d, 0xf3, 0xeb, 0x66, 0xd2, 0x5f, 0x68, 0x64, 0xb9, 0x1b, 0xba, 0x78, 0x5f,
+	0x5b, 0xc7, 0x5e, 0xe0, 0x86, 0xc7, 0x96, 0xd0, 0x9f, 0xc1, 0x05, 0xfb, 0x08, 0xee, 0x6c, 0xd3,
+	0x3e, 0xde, 0x0d, 0x5d, 0xb8, 0x39, 0x3f, 0x40, 0x16, 0xee, 0xec, 0xc5, 0x6e, 0x09, 0x29, 0x1a,
+	0xe5, 0x32, 0x9c, 0xaf, 0x1c, 0xdc, 0xca, 0x63, 0x5e, 0xcc, 0x11, 0x1f, 0xf4, 0x53, 0x8d, 0xac,
+	0x65, 0xc7, 0xc4, 0x39, 0x8a, 0x40, 0x9b, 0x75, 0x1c, 0x79, 0x31, 0x17, 0xfa, 0xb3, 0x28, 0xe6,
+	0xfb, 0x50, 0x7a, 0xd3, 0x84, 0xcf, 0xf8, 0x0f, 0x90, 0x4e, 0x24, 0xbb, 0xa5, 0x9c, 0x9a, 0x12,
+	0xa7, 0x1c, 0x9e, 0x2d, 0xe5, 0xec, 0x68, 0x5b, 0xe6, 0x24, 0x4f, 0x50, 0xc4, 0xf2, 0xdc, 0x6e,
+	0xc1, 0xbb, 0x4e, 0xdf, 0x18, 0x16, 0xb1, 0x8c, 0x68, 0x00, 0x5e, 0x1c, 0x7e, 0x15, 0x34, 0xcc,
+	0x92, 0x0d, 0xf5, 0xc9, 0x12, 0xbe, 0xb7, 0x2d, 0xa8, 0x05, 0x56, 0x5a, 0x5f, 0x19, 0xd6, 0xd7,
+	0x1b, 0x79, 0x7d, 0xad, 0x03, 0x3f, 0x2c, 0xb2, 0xf8, 0x04, 0x39, 0x28, 0x61, 0xc5, 0xca, 0x96,
+	0x61, 0xc3, 0x1c, 0xb1, 0xa3, 0x9f, 0x6b, 0x64, 0x19, 0x53, 0x08, 0x9f, 0xeb, 0x56, 0xfa, 0x5e,
+	0xd7, 0x2b, 0x18, 0x6f, 0x05, 0x9e, 0x3b, 0xdb, 0x61, 0xaf, 0x6f, 0x02, 0xb7, 0x8b, 0x54, 0xfd,
+	0x1e, 0x34, 0x8c, 0x4e, 0x19, 0x4c, 0x24, 0xab, 0x15, 0x69, 0xa4, 0xe0, 0xca, 0x32, 0x8a, 0xd8,
+	0x0e, 0x5c, 0x3b, 0x72, 0xe1, 0xfe, 0x9f, 0xcd, 0x07, 0xe6, 0xa8, 0x23, 0xfa, 0x47, 0x90, 0x63,
+	0x43, 0x01, 0xe5, 0x81, 0xf0, 0x62, 0xef, 0x01, 0xac, 0xa8, 0xfe, 0x1c, 0x2e, 0xe7, 0x09, 0x74,
+	0xaf, 0xdb, 0xb6, 0xe0, 0xcd, 0x9c, 0x6b, 0x60, 0xf7, 0xea, 0x94, 0xa1, 0x44, 0xb2, 0xb5, 0x54,
+	0x4c, 0x19, 0x87, 0x1e, 0x68, 0xcc, 0x76, 0x1c, 0x82, 0x9e, 0x75, 0x24, 0x88, 0x39, 0x62, 0x23,
+	0xe8, 0x1f, 0x34, 0xb2, 0xd4, 0x0a, 0x7d, 0x3f, 0x3c, 0xb6, 0x3e, 0x3e, 0x0a, 0x1c, 0x68, 0x47,
+	0x84, 0x6e, 0x0c, 0x55, 0x7e, 0x2f, 0x07, 0xdf, 0x15, 0x3b, 0x5e, 0x24, 0x40, 0xe5, 0xc7, 0x65,
+	0xa8, 0x50, 0x39, 0x82, 0xa3, 0xca, 0x51, 0xdb, 0x71, 0x08, 0x54, 0x8e, 0x04, 0x31, 0xaf, 0xa7,
+	0x8a, 0x0a, 0x98, 0xde, 0x27, 0x8b, 0x90, 0x51, 0xc3, 0xea, 0xa0, 0x3f, 0x8f, 0x12, 0xe1, 0x15,
+	0xb8, 0x00, 0x4c, 0x71, 0xae, 0x13, 0xc9, 0x56, 0xd2, 0xcb, 0x4f, 0x45, 0x0d, 0xb3, 0x6c, 0x85,
+	0x0e, 0x79, 0xe0, 0x2a, 0x0e, 0xab, 0x8a, 0x43, 0x1e, 0xb8, 0x13, 0x1c, 0xaa, 0x28, 0x38, 0x54,
+	0xc7, 0x50, 0x04, 0x51, 0xe1, 0x09, 0x74, 0xa3, 0x42, 0xbf, 0x85, 0xde, 0xb0, 0x08, 0x02, 0xfc,
+	0x23, 0x44, 0x8b, 0x22, 0x38, 0x84, 0x0c, 0x53, 0xe1, 0xd1, 0x09, 0xa8, 0xca, 0x9c, 0xbc, 0xa0,
+	0x38, 0xe1, 0x81, 0x3b, 0xea, 0xa4, 0x80, 0xc0, 0x49, 0x31, 0x80, 0xc6, 0x1e, 0xe7, 0xc3, 0xdd,
+	0x17, 0xf3, 0x48, 0x7f, 0x11, 0x7b, 0xd0, 0x95, 0xfc, 0xc4, 0xa1, 0x55, 0x03, 0xa9, 0x7a, 0x2d,
+	0x6f, 0x7c, 0x4f, 0x86, 0x60, 0x22, 0xd9, 0x32, 0xfa, 0x57, 0x30, 0xc3, 0x54, 0x2d, 0xe8, 0x21,
+	0x99, 0x8b, 0xb8, 0xed, 0x5a, 0x61, 0xe0, 0xf7, 0xf5, 0x3f, 0x37, 0x50, 0xe5, 0xee, 0x85, 0x64,
+	0x74, 0x87, 0xf7, 0x22, 0xee, 0xd8, 0x31, 0x77, 0x4d, 0x6e, 0xbb, 0xf7, 0x03, 0xbf, 0x3f, 0x90,
+	0x4c, 0x7b, 0xb5, 0xf8, 0xc7, 0x27, 0x0a, 0xb1, 0x59, 0x7f, 0x25, 0xec, 0x7a, 0x70, 0x73, 0xc6,
+	0x7d, 0xfc, 0xc7, 0x67, 0x0c, 0xd5, 0x35, 0x73, 0x36, 0xca, 0x1c, 0xd0, 0x9f, 0x91, 0xe5, 0x52,
+	0x07, 0x8f, 0xb7, 0xd9, 0x5f, 0x1a, 0xf8, 0xa2, 0x7a, 0xef, 0x42, 0x32, 0x7d, 0x18, 0x74, 0x77,
+	0xd8, 0x87, 0xef, 0x39, 0x71, 0x1e, 0x7a, 0x63, 0xb4, 0x8d, 0xdf, 0x73, 0x62, 0x45, 0x81, 0xae,
+	0x99, 0x8b, 0x65, 0x92, 0xfe, 0x98, 0x5c, 0x4d, 0xbb, 0x17, 0xa1, 0x7f, 0xd9, 0xc0, 0xca, 0xfb,
+	0x2d, 0xb8, 0x06, 0x86, 0x81, 0xd2, 0xae, 0x54, 0x94, 0x3f, 0x2e, 0x9b, 0xa2, 0xb8, 0xce, 0xca,
+	0xad, 0xae, 0x99, 0xb9, 0x3f, 0x7a, 0x48, 0x16, 0xb1, 0xaf, 0x1b, 0xe6, 0xdd, 0x5f, 0xd3, 0xf5,
+	0xdb, 0xbe, 0x90, 0xec, 0xe6, 0x30, 0x42, 0xd3, 0xb1, 0x83, 0x22, 0xb9, 0xf2, 0x38, 0xcf, 0x16,
+	0x5d, 0x5d, 0x41, 0x95, 0x3f, 0x64, 0xa1, 0xc4, 0x19, 0x9f, 0xcd, 0x90, 0x79, 0x65, 0xbb, 0xe9,
+	0x47, 0xe4, 0x2a, 0x0f, 0xe2, 0xc8, 0xe3, 0x42, 0xd7, 0xf0, 0x3f, 0x10, 0x7d, 0x42, 0x52, 0xbc,
+	0x17, 0xc4, 0x51, 0xbf, 0xfe, 0x62, 0xfe, 0xd7, 0x47, 0x36, 0xa1, 0xe8, 0x79, 0x61, 0x8c, 0xdb,
+	0x76, 0x19, 0x7f, 0x99, 0xb9, 0x01, 0xfd, 0x5d, 0x76, 0x79, 0x09, 0x2f, 0x68, 0xfb, 0xdc, 0x42,
+	0xd6, 0x12, 0xf0, 0xe8, 0x9b, 0xc6, 0x25, 0x6c, 0x41, 0x5f, 0xd4, 0xb5, 0x4f, 0x9a, 0xc8, 0x63,
+	0x94, 0xa6, 0xfa, 0xf2, 0x1b, 0xa7, 0x4a, 0x7d, 0xdf, 0xd6, 0x1b, 0xca, 0x23, 0x62, 0x82, 0x1f,
+	0x78, 0x00, 0x82, 0x95, 0x39, 0x81, 0xa3, 0x0f, 0xc9, 0x22, 0x48, 0x8b, 0xc3, 0x18, 0x1a, 0x68,
+	0xd0, 0x34, 0x83, 0x9a, 0xf6, 0xb3, 0xfe, 0x73, 0x1f, 0x88, 0x4c, 0xcd, 0x73, 0xb9, 0x9a, 0x02,
+	0x54, 0x74, 0xbc, 0x71, 0xe7, 0xad, 0x37, 0x15, 0x1d, 0xa5, 0xb9, 0xa0, 0x00, 0x78, 0xb3, 0x84,
+	0x1a, 0xbf, 0xd7, 0xc8, 0xd2, 0xe8, 0xf2, 0xc2, 0x73, 0xa3, 0x0b, 0xef, 0xf1, 0xec, 0x6f, 0xc4,
+	0x6f, 0xc0, 0xdb, 0x02, 0x01, 0xa5, 0x4f, 0x8a, 0x9d, 0x4e, 0xf1, 0xd2, 0x26, 0xc3, 0xa1, 0x99,
+	0x1a, 0xd2, 0x06, 0xb9, 0x02, 0x0f, 0x77, 0x2f, 0xc6, 0xf5, 0x9d, 0xad, 0x6f, 0x62, 0x7f, 0x88,
+	0x48, 0x71, 0x84, 0xd3, 0x61, 0xe1, 0x65, 0x5e, 0x19, 0x9b, 0x99, 0x6d, 0xfd, 0xde, 0xa3, 0xaf,
+	0x36, 0xa6, 0xce, 0xbf, 0xda, 0x98, 0x7a, 0x74, 0xb1, 0xa1, 0x9d, 0x5f, 0x6c, 0x68, 0xbf, 0x7e,
+	0xbc, 0x31, 0xf5, 0xc5, 0xe3, 0x0d, 0xed, 0xfc, 0xf1, 0xc6, 0xd4, 0xbf, 0x1f, 0x6f, 0x4c, 0x7d,
+	0xf8, 0xd2, 0xff, 0xf0, 0xaf, 0x6f, 0x9a, 0x47, 0x07, 0x57, 0xf0, 0xdf, 0xdf, 0xd7, 0xff, 0x1b,
+	0x00, 0x00, 0xff, 0xff, 0x68, 0x19, 0x36, 0x63, 0x1b, 0x18, 0x00, 0x00,
 }
 
 func (m *FolderDeviceConfiguration) Marshal() (dAtA []byte, err error) {
@@ -510,6 +514,14 @@ func (m *FolderConfiguration) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 		dAtA[i] = 0xb2
 		i--
 		dAtA[i] = 0xc0
+	}
+	if m.FSWatcherTimeoutS != 0 {
+		i -= 8
+		encoding_binary.LittleEndian.PutUint64(dAtA[i:], uint64(math.Float64bits(float64(m.FSWatcherTimeoutS))))
+		i--
+		dAtA[i] = 0x2
+		i--
+		dAtA[i] = 0xc1
 	}
 	{
 		size, err := m.XattrFilter.MarshalToSizedBuffer(dAtA[:i])
@@ -1107,6 +1119,9 @@ func (m *FolderConfiguration) ProtoSize() (n int) {
 	}
 	l = m.XattrFilter.ProtoSize()
 	n += 2 + l + sovFolderconfiguration(uint64(l))
+	if m.FSWatcherTimeoutS != 0 {
+		n += 10
+	}
 	if m.DeprecatedReadOnly {
 		n += 4
 	}
@@ -2199,6 +2214,17 @@ func (m *FolderConfiguration) Unmarshal(dAtA []byte) error {
 				return err
 			}
 			iNdEx = postIndex
+		case 40:
+			if wireType != 1 {
+				return fmt.Errorf("proto: wrong wireType = %d for field FSWatcherTimeoutS", wireType)
+			}
+			var v uint64
+			if (iNdEx + 8) > l {
+				return io.ErrUnexpectedEOF
+			}
+			v = uint64(encoding_binary.LittleEndian.Uint64(dAtA[iNdEx:]))
+			iNdEx += 8
+			m.FSWatcherTimeoutS = float64(math.Float64frombits(v))
 		case 9000:
 			if wireType != 0 {
 				return fmt.Errorf("proto: wrong wireType = %d for field DeprecatedReadOnly", wireType)
